lib: explicit cast to unsigned

+++ lib/printf/vfprintf.c
@@ -453,7 +453,7 @@ reswitch:	switch (ch) {
 			if (flags & INTMAX_SIZE)
 				ujval = SJARG();
 			else
-				ulval = SARG();
+				ulval = (u_long)SARG();
 
 			if (printfrr_ext_char(fmt[0])) {
 				struct printfrr_eargs ea = {
@@ -483,7 +483,7 @@ reswitch:	switch (ch) {
 				}
 			} else {
 				if ((long)ulval < 0) {
-					ulval = -ulval;
+					ulval = (u_long)-ulval;
 					sign = '-';
 				}
 			}
